board/reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
pymongo==3.11.0
requests==2.23.0
Flask-PyMongo==2.3.0
pylint==2.4.0
jaeger-client==4.3.0
opentelemetry-api==1.6.2
opentelemetry-sdk==1.6.2
opentelemetry-instrumentation==0.25b2
opentelemetry-instrumentation-flask==0.25b2
opentelemetry-instrumentation-requests==0.25b2
opentelemetry-exporter-jaeger-thrift==1.6.2
gunicorn
prometheus_client
prometheus_flask_exporter